Jotunn Damage

+++ RobberJotunn/Assets/Scripts/JotunHealth.cs
@@ -15,7 +15,7 @@ public class JotunHealth : MonoBehaviour
     }
 
     // change health
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
     }
 }
